+++ src/frontend/app/shared/components/list/list-types/cf-users/cf-user-list-helpers.ts
@@ -1,0 +1,13 @@
+import { CfUser } from '../../../../../store/types/user.types';
+
+export const userListUserVisibleKey = 'showUsers';
+
+export enum UserListUsersVisible {
+  ALL = 'all',
+  WITH_ROLE = 'withRole',
+  NO_ROLE = 'noRole'
+}
+
+export const userHasRole = (user: CfUser, roleProperty: string): boolean => {
+  return !!user[roleProperty] && !!user[roleProperty].length;
+};
